extract GroupAdmin Controller
